Bully Follow Player Refactor - Update Variables in Player Controller
- Bully Follows player and steel candy and stop following for a while
- Change candy float to int variables in PlayerController

+++ Assets/Scripts/Player/PlayerController.cs
@@ -41,13 +41,13 @@ public class PlayerController : MonoBehaviour
         rb.MovePosition(rb.position + displacement);
     }
 
-    public void AddCandy(float amount)
+    public void AddCandy(int amount)
     {
         currentCandy += amount;
         currentCandy = Mathf.Clamp(currentCandy, 0, maxCandy);
     }
 
-    public void LoseCandy(float amount)
+    public void LoseCandy(int amount)
     {
         currentCandy -= amount;
         currentCandy = Mathf.Clamp(currentCandy, 0, maxCandy);
